% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{xSelectCells}
\alias{xSelectCells}
\title{Select cells from clusters}
\arguments{
\item{input_seurat_obj}{a seurat obj with umap reduction}
}
\description{
A light weight shiny app tool to enable manually selection of cells from seurat object based on its umap reduction
}
\details{
xSelectCells
}
\examples{
xSelectCells(seurat_obj)
}
\author{
Ashley(X) Zhao
}
